improve responsiveness of tables
