+++ application/vue/src/widgets/EditableTable/widgets/EditableCell.tsx
@@ -50,10 +50,10 @@ const EditableCell = <RecordType extends AnyObject = AnyObject>({
     }
   }, [editing]);
 
-  const save = async () => {
+  const save = async (keepEditing?: boolean) => {
     try {
       const values = await form.validateFields();
-      toggleEdit();
+      !keepEditing && toggleEdit();
       handleSave(record, values);
     } catch {
       /* empty */
